Support display-only information in place of a suffix
Group headings are also display-only information and raw strings
can be used anywhere a suffix can appear, but this class makes it
possible to display arbitrary information anywhere *and* to style
it in any way one fancies.

By inheriting from `transient-suffix' (instead of `transient-child')
we avoid having to implement many new methods, but semantically this
is a bit questionable.  Leaving the `command' and `key' slots
undefined, means that we have to use `slot-boundp' in a few places,
but on the other hand that helps dealing the semantic inconsistency.

Closes #226.