transient--make-predicate-map: Fix handling of transient-suffix
The value of the prefix's `transient-suffix' slot has to be handled
differently for different types of suffixes.  It makes little sense
to use a predicate as value, and a boolean should be used instead.

In the past a predicate had to be used, so we have no choice but to
continue to support that, and for uniform suffixes it may still make
sense to use a predicate.  Also substitute some known predicates based
on the suffix type.

- Infixes must disregard the default transient behavior specified
  for all suffixes of the current prefix.

  For a select few infixes a different behavior might make sense.
  If so, then that must be specified for those infixes individually.

- Sub-prefixes must honor the default transient behavior specified for
  all suffixes of the current outer prefix, but to do that they have
  to use predicates different from those used for regular suffixes.

The default transient behavior is interpreted more restrictively than
the transient behavior explicitly specified for a single suffix, to
limit surprising behavior but still allow funky behavior if one really
wants that.