Account for t as transient-suffix for nested prefixes
When the prefix's `transient-suffix' is t, and the current prefix
is a sub-prefix, then suffixes have to use `transient--do-return',
to exit the sub-prefix but not the outer prefix.

Previously this only worked if the value of that slot was that
predicate itself, but that must be avoided because it is only
suitable for suffixes, not infixes and sub-prefixes.