transient--stack-push: Preserve value of transient-suffix slot
Given a prefix nesting that is three or more levels deep, where
exiting any of the sub-levels is expected to return to the outer
level, that failed except for the innermost level, because the
value of a (sub-)prefix's `transient-suffix' slot is set to t
based on settings in the outer prefix, and technically there is no
outer prefix *active* when we return to an outer level (instead it
lives in the stack, from where we pop it).

We can pass the slot's value between incarnations of the sub-prefix's
object, by including it in the data stored in the stack.