transient--unset-incompatible: Remove unnecessary variable
When a variable is set to a non-nil value, then the value of all
incompatible variables has to be set to nil.  Setting these other
variables to nil is unproblematic, we don't have to do anything
special; the mere fact that we are setting to nil, not non-nil,
prevents a cascading effect.