transient-infix-set(argument): Limit to other arguments
When checking whether other suffixes are incompatible (and this
method therefore has to disabled them) ignore all suffixes that are
not `transient-argument' objects.  Previously we instead checked
whether they have an `argument' slot.  (Which all `transient-infix'
do, for dubious reasons).

Since we only check for conflicts when an *argument* is enabled, it
makes sense if the other side of the comparison is also limited to
arguments.  Besides, we cannot know what a potential conflict with
a non-argument would look like, and don't now what would have to be
done to detect it.