transient-infix-set(argument): Fix disabling incompatible options
When `argument' is unbound, fall back to `argument-format' instead of
`argument-regexp'.  The format can easily be used both as the matcher
and the match candidate, while the regexp is only suitable as matcher.

Fallback from `argument' to `argument-format', not only for the
argument that is being enabled, but also for the arguments that might
have to be disabled in response.