transient-format: No longer apply face for inapt bindings here
Apply a face individually for the key and the description.  For the
key, always use `transient-inapt-suffix'.  For the description honor
the new `inapt-face' slot.

Some suffixes are inapt because they represent the current state and
invoking them therefore wouldn't makes sense.  In that case we likely
want to highlight them, instead of dimming them.