transient--do-replace: Replace with two variants
Rename the existing `transient--do-replace' to `transient--do-stack',
because that's what it really does.

Repurpose the name `transient--do-replace' for a new variants that
actually replaces the active transient, i.e., which does not place
the active transient on the stack, but replaces it completely.

Changing what `transient--do-replace' does, shouldn't be a problem
because in definitions of individual transient prefix commands and
their sub-prefixes, this pre-command should never have been used
by name anyway.