Fix/improve some faces
Derive `transient-argument' from `font-lock-string-face' instead of
`font-lock-warning-face'.  Enabling an argument doesn't warrant a
warning.  `transient-value' also derives from `font-lock-string-face'
and it makes sense for these two to look the same.

Derive `transient-active-infix' from `highlight' instead of
`secondary-selection'.  That looks better with the light default
theme.  With the dark default theme both look okay.

The goal of `transient-unreachable-key' is to dim the key, which we
hope to achieve by inheriting from `shadow', but for that to work,
that face must take precedence, which means it must come first in
the list of inherited faces.