Simplify how colors are used to indicate transient behavior
The old implementation mimicked how Hydra uses them.  It's was a neat
idea, but in practice I was never able to remember what the difference
between a pink and a amaranth prefix/hydra is, for example.

So I'm giving up on encoding the separate behaviors of suffixes and
non-suffixes, using a single color.  Instead the behavior of each
individual suffix is visualized by coloring its key binding (as
before), and by using the separator line to visualize the behavior
common to all non-suffixes (without additionally encoding the default
suffix behavior by blending colors).

Also change the colors used for "exit" and "stay" behavior.  IMO it
makes more sense if "red" means "exit" instead of "stay", but of
course you are welcome to disagree.  Its much less mind-bending to do
that now, because the new faces use names that describe their behavior
instead of an arbitrary color encoding.  Previously one would have had
to make `transient-red' "green" and `transient-blue' "red" to achieve
the new default behavior, for example.

Finally, now that this has been simplified, enable it by default.