Declare version with "Version" rather than "Package-Version"
Emacs builtin packages are supposed to specify their version using the header
"Version" for their version to be correctly stored in the variable
package--builtins.

This is because the function finder-compile-keywords, responsible for setting
this variable, parses the source files looking for this exact header.

This does not break package, since the function package-buffer-info looks for
both Package-Version and Version.