transient-format(around:suffix): Don't attempt to highlight full line
This only worked if the user sets the `extend' attribute of the
`transient-active-infix' face to t (or uses an Emacs version, that
does not contain this intentional regression).

This assumes that the infix is displayed in the only or last column,
which is very likely to be the case, but not guaranteed.  If we can
figure out how to make it known here whether that is the case, then
we can consider bringing whole-line highlighting back.