Improve looks of debugging faces
Most importantly give them distinct looks, and use colors appropriate
to the severity of the potential issue.  I.e., mismatched keys should
be avoided if at all possible, while adding a shorthand-like binding
for an argument that doesn't exist as a shorthand, cannot be avoided,
and hiding come suffixes from users is also something that is often
done intentionally.