Define box line-width more carefully
Before Emacs 28, the value had to be a number.  In Emacs 30 it can
be a cons-cell of two numbers, but the manual says that it can still
be a number.  But that doesn't seem to work for all Emacs releases
in between, at least not when using a negative number.

Try to cater to the preferences of various Emacs releases.