transient--delete-window: Get transient buffer by name
We try to dedicate the transient window to the transient buffer, but
that does not actually guarantee that it is not used to displayed
another buffer.  Normally displaying another buffer in this window
would be fatal, but if it happens when displaying the *Help* buffer,
then that works out okay, as long as we avoid immediately killing
that buffer again.

Closes #271.