transient--make-predicate-map: Fix detecting outer prefix
Since [1: 5f2cfc9f] we try to detect whether there is an outer
prefix, that non-transient suffixes could return to, by checking
if `transient--current-prefix' is non-nil.  That works then the
prefix is created, but that variable is also always non-nil after
a transient suffix is invoked.  Instead check if `transient--stack'
is non-nil.

1: 2023-11-29 5f2cfc9f73e756fb337df70909fcd3d0bce56911
   transient--make-predicate-map: Only return if there is a parent