transient--parse-suffix: Avoid overly long symbol names
We already fall back to making the suffix's key part of the name of
the command, when its description is a lambda.  Do the same if it is
a named function, and if it is a string, then only use it if it is
less than 16 characters long.

(As a side-note, using the description has the advantage that more
meaningful symbols appear in backtraces.  While falling back to using
the key, is less meaningful (and users may change the key, leading to
a mismatch), it is still much more useful than using `gensym'.)