transient-parse-{suffix,suffixes}: Assert PREFIX is a non-nil symbol
This symbol is used in the names of suffixes that are defined inline.

The idea was that callers would provide the name of the corresponding
prefix as a quoted symbol.  We do not check whether it actually is
prefix command, much less the correct prefix command, because the same
setup function may be shared between different prefix commands.

In #277 a user provided an example, that instead used unquoted
`transient--prefix', which is wrong since this is an object not a
symbol, resulting in extremely long names for.

(oref transient--prefix command) would work though.  (Maybe we should
start falling back to that if PREFIX is nil.  For now, let's not.)