Document that :class might be needed if :setup-children is used
It was undefined which class would be used if none was specified
explicitly and `:setup-children' is used.  Using `transient-column'
makes sense because, at least for a flat list of suffixes, that is
the default for explicitly specified suffixes too.

[1: df36bc87] accidentally changed the undefined default behavior to
use `transient-columns' instead.  This commit reverts that, but it is
still considered undefined behavior.  We now display a warning if this
is encountered, and that will eventually be turned into an error.

(Note that while not particularly useful, it is allowed to define a
group with neither explicit children nor `:setup-children' (it will
be silently omitted).  At least for now, we keep it that way.)

Closes #277.

1: 2024-02-01 df36bc8745520f4580c8bc9260d57aa53878a053
   transient--parse-group: Detect columns beginning with included group