Revert "transient--show: Fix only window height"
This reverts commit 03997dcba8d5301b6eb26754773297f23ba3eca2.

This avoids an issue for the `transient-posframe' package, mentioned
at https://github.com/magit/transient/commit/03997dcba8d5301b6eb2675.

It is unknown why fixing only the height caused the height to be
reduced, but since we fixed both the height and width for a long time,
without *that* causing any issues, just go back to doing that.