transient--get-description: New function
As a result, `transient-format-description(transient-group)' no longer
has to call the next method.

As a result of that in turn, `transient--get-face' no longer gets called
for objects that lack the `face' slot.  Remove the `slot-exists-p' check
and restrict to `transient-suffix', which all have that slot.  Restrict
type of `transient--get-face' as well.

We do all this, so that we can, in a later commit, add a `face' slot to
`transient-group', but use it for a different purpose.