transient-format-description(suffix): Move fontification to around method
Like for other fontification work already performed in the around
method, this means that specialized primary methods of derived
classes, do not have to duplicate the code necessary to do this.

As a result, primary methods can no longer prevent fontification by
simply not doing it, but that is not a problem because fontification
can be prevented by leaving the `face' slot unbound, or by setting
it to nil.