transient--refresh-transient: Preserve value
The value of the prefix's `value' slot is used during the setup of its
suffixes.  After that it is not updated, when the value of a suffix
changes (and `transient-get-value' has to be used instead).  Since
`transient--refresh-transient' redoes part of the setup, it has to make
sure it uses the updated value.

As a special-case, if a suffix actually does set the prefix's `value',
then it has to signal that, so we can avoid discarding it here.  Add
a new function `transient-prefix-set' to handle that situation.