transient--post-command: Clear transient-current-* variables here
These variables are only intended to be accessible by the command that
was invoked as a suffix of a prefix.  It we only clear their values when
we exit the prefix (but not when a non-exiting suffix is done), then the
values remain accessible in timers and while the prefix is suspended
(e.g., while using the minibuffer).