transient--post-command: Only clear variables on real post-command
In [1: 0e0ece32] we started clearing `transient-current-*' variables
here instead of in `transient--post-exit', to make sure their values
are nil in timers and while using the minibuffer.

The latter was achieved by clearing even during the premature call to
`transient--post-command', which occurs when the minibuffer is entered,
i.e., when suspending the transient.  Unfortunately transient state
cannot be resumed until the outer command really is done, and as a
result these variables remained cleared for the current command after
any use of the minibuffer.

This obviously can be fixed by only clearing on the real `post-command'
run, but that of course means that these variables are once more
accessible during minibuffer use.  There does not appear to be a way
around that, but luckily it is unlikely that that would actually lead
to any unexpected behavior.

Closes #292.

1: 2024-06-26 0e0ece32362bb5eed430328583018a35f44d3c7d
   transient--post-command: Clear transient-current-* variables here