transient--parse-suffix: Ensure generated names are unique
Two suffixes may share a common description.  They may even share the
description and the key binding, without also sharing the definition
(though that only makes sense if they don't also share the predicate).

To ensure the names are unique, include `gensym-counter' in the name.

A drawback of that is, that if the prefix definition is evaluated
multiple times, we also get multiple definitions for each suffix that
gets such a generated name.

Give up on including the description in the name and remove whitespace
from the key before including it in the name.