transient--describe-function: Refactor setting help buffer/window
Bind `help-window-select' to select the window used to display the help
buffer.  Use `temp-buffer-window-setup-hook' to get our hands at that
buffer, despite `describe-function' (and the lower-level functions it
uses) neither returning the used buffer nor making it current.

`transient-show-help' expects the used buffer to be current after
calling functions such as this one, so that it can finish its own setup
of that buffer.