transient--parse-suffix: Fix regression breaking :setup-children
In [1: fb88f803] we wrongly concluded that the only type of function
left to be handled by this branch are `lambda' expressions.  While that
is true during macro-expansion, this function may also be called from a
prefix's `:setup-children' function, in which case we also have to deal
with byte-code function objects and compiled functions.

So go back to handle everything for which `commandp' returns non-nil,
except for strings, which Transient does not support as commands.

Because none of command types supported here have to be quoted, we can
keep the part of [1: fb88f803], which stopped using `macroexp-quote'.

Closes #313.
Closes tarsius/notmuch-transient#2.

1: 2024-09-02 fb88f803e85fa0b9e1319e90b5e71020b6543d94
   transient--parse-suffix: Clarify clause, removing dead code