Ensure transient-current-* variables are unset on every exit
In the simple case, when `transient--post-command' determines that
the transient is being exited, it calls `transient--post-exit' to
unset internal state but takes care to unset `transient-current-*'
itself.

When the minibuffer is involved and for other recursive edits and
if an unexpected error occurs, it cannot take care of the latter.

In those cases we can do it in `transient--post-exit' instead, but
until now we didn't.  Do so if COMMAND is non-nil, which in the
normal case it is not.

Closes #323.