transient-suffix-object: Restore fallback behavior during setup
In [1: 1b7f9be7] we started to signal an error if there are multiple
bindings for the command and we cannot determine the correct suffix
object based on the key binding.

When COMMAND is non-nil, then this function is not being called on
behalf of `this-command'.  In that case `this-command-keys' is also
meaningless, and when COMMAND has multiple bindings, then COMMAND
cannot be mapped to a single suffix object.  Go back to returning
the first object.

Closes #325.

1: 2024-11-02 1b7f9be7b9fbc8acb5d71f028b19e1f78d9dd090
   transient-suffix-object: Drop faulty fallback behavior