transient-hide-during-minibuffer-read: Change default to t
That way the menu is consistently hidden during all minibuffer usage
by suffix commands, *including* when setting infix arguments.

Beside the increased consistency, this is beneficial for large menus
(such as `magit-log') in combination with completion setups that
immediately show a set of completion candidates.  In such cases there
could be very little space left to display anything beside the menu
and the completion candidates.

Additionally this helps to avoid an incompatibility when displaying the
menu using the full frame height.  See the next commit for details on
that.