transient--show: Fix window height for more setups
It is desirable to set `window-size-fixed' to t in the transient menu,
as we always used to do, until very recently.

We cannot do this when the transient window uses the full height of the
frame, the minibuffer may be used while the menu remains displayed, and
the height of the minibuffer has to be increased so that the completion
framework can display completion candidates.

If `transient-hide-during-minibuffer-read' is t [1], then that cannot
happen, so we fix the size in that case again.

[1] Which is the default, starting with the parent of this commit.

Re #532.