transient--stack-push: Preserve the value
Since [1: 179545a6] we re-initialize suffixes when returning to the
outer prefix, that also resulted in the value of the prefix and the
values of the suffixes being re-initialized.

We can prevent that by storing the value on the prefix stack, as we
are already doing for the scope.

The value of a prefix's `value' slot is not updated when the user
changes the value of a suffix, so we cannot use the current value
of that.  Instead we have to collect the up-to-date value from the
suffixes.

Closes #333.

1: 2024-11-13 179545a6e29899aa89cbd77962a1c882f45927ca
   Re-initialize suffixes when returning to :refresh-suffixes prefix