transient-default-value: Remove transient-suffix method
In [1: eff2245b] we stopped defining default methods using
`cl-defgeneric' and added explicit default methods were this seemed
necessary, including for this function.

Turns out it is not necessary in this case and even misleading, because
for `transient-suffix' objects, `transient-init-value' does not actually
use `transient-default-value', only for prefix objects.

In the event that someone implements a `transient-init-value' method
and that uses `transient-default-value', but fails to implement the
corresponding method for the latter, that now results in an informative
error message (instead of nil being used silently).

(In the future we might want to start using `transient-default-value'
for `transient-suffix' objects as well, but that will/would require
changes to existing `transient-init-value' methods and thus shouldn't
be done hastily.)

1: 2024-12-05 eff2245b73ceb471afb0d74c78bf57994767d5b0
   Never use cl-defgeneric to define the default method