Restore ability to change a prefix's level
In [1: 1b7f9be7] we intentionally started signaling an error in
`transient-suffix-object' if no object can be determined, to learn
whether there actually are any callers for which it actually make
sense to return nil instead.

Now we know about one, but I am still not sure whether that should
be a valid use-case, so we keep the error for now, but ignore the
error in this one case.

Closes #337.

1: 2024-11-02 1b7f9be7b9fbc8acb5d71f028b19e1f78d9dd090
   transient-suffix-object: Drop faulty fallback behavior