Fix setting up return to outer prefix
We need a separate prefix slot, `return', to store the return behavior;
re-purposing the `transient-suffix' slot, as we did before, was
ambiguous.

When the `transient-suffix' prefix slot is t, that means that suffixes,
that do not specify their own transient behavior, should be called
without the prefix being exited.

In [1: 784887b7] we tried to overload that meaning of that value in the
case of sub-prefixes:

> When the prefix's `transient-suffix' is t, and the current prefix
> is a sub-prefix, then suffixes have to use `transient--do-return',
> to exit the sub-prefix but not the outer prefix.

If a suffixes `transient' slot is nil, that works, but merely because
the ambiguous value t of `transient-suffix' does not enter the picture.

However if a suffix leaves `transient' undefined, then we relied on the
same boolean to encode the three possible behaviors `call', `exit' and
`return'.  I.e., it was not possible to chose between `exit' and
`return', and we always used `exit'.

Closes #352.

1: 2023-10-31 784887b78160e0c820b44924d12b2e2e0bf90ef0
   Account for t as transient-suffix for nested prefixes