Replace transient--do-return when there is no outer prefix
A suffix may explicitly be configured to return to the outer prefix
even if there is no outer prefix.  This may happen, for example, if
a prefix is invoked directly, even though it usually is invoked via
an outer prefix.

`transient--do-return' itself already takes care to fall back to
`transient--do-exit' behavior in this situation, but that does not
affect to color used for the binding.  By detecting this situation
in `transient--make-predicate-map', we can also correct the color.