manual: Update how to prevent C-g from returning to the parent prefix
Nowadays we have more explicit pre-commands for prefixes we can
choose from, and don't have to rely on `transient' being unbound.
The documented behavior no longer worked; nil has been a shorthand
for `stack' (itself a shorthand for `transient--do-stack') for a
long time, and leaving the slot unbound meant that the default was
used, which is also `transient--do-stack'.

Now `replace' prevents even "C-g" from returning to the parent, and
`recurse' makes all suffixes of the sub-prefix default to returning
to the parent, not just "C-g".

Closes #356.