transient--suspend-override: Assert transient--window is still live
It should be, but something else might mess with it.

Helm's completing-read function does mess with it.  It *appears*
(I haven't investigated further) that it tries to use our window,
but because we prevent that, it instead deletes that window and
creates its own window.  It does however bring back our buffer,
when done.  Still, we cannot assume here that our window is live.

Closes the second coming of #656.