Split "Draw" section in two
Replace the "Draw" section with two new sections "Display" and "Format".
The two tasks are closely related (when it comes to the part that deals
with the window), but they are still distinct.

`transient--show', located in the new "Display" section, still concerns
itself with formatting/inserting the menu.  A subsequent commit tackles
that.