transient--wrap-command: Use add-function and remove-function
While suffixes are guaranteed to be function-bound symbols, the same
isn't necessarily true of non-suffixes.  Non-suffixes may also use
the minibuffer or signal an error, so they have to be wrapped as well.

Furthermore when `advice-add' is used on a function whose doc-string
is stored in its `function-documentation' property, then the original
doc-string is lost once `advice-remove' is used.

Re debbugs#76528.