transient--wrap-command: Advise anonymous commands via this-command
In [1: aea527ed] we attempted to also wrap anonymous non-suffixes.
The chosen approach does not work because `cmd' is a local variable,
which has no meaning and by definition cannot be accessed outside of
`transient--wrap-command'.

For named commands we continue to advise via the function stored in
the respective symbol.  (An alternative would have been to deal with
named function the same way we deal with anonymous functions, as
described below, but that would mess with commands that internally
look at `this-command' and expect to find a symbol.  It is safer to
stick to the current behavior, instead of switching to another, for
consistency reasons.)

For anonymous commands we use `this-command' instead.  That works
because it is the global variable, whic the command loop uses to
lookup the command.

Do not try to remove the advice from `this-command'.  Doing so is
unnecessary because the command loop takes care of it, and if the
value of `this-command' is changed after we advised it (likely by
the command itself), then doing it explicitly would fail.  Do not
do it explicitly just because, as that would have a greater change
of confusing me later on.

1: 2025-02-26 aea527ed938de32a96ec138b844b5d0c9643dada
   transient--wrap-command: Use add-function and remove-function

Suggested-by: Stefan Monnier <monnier@iro.umontreal.ca>