Require all key bindings to be specified as key descriptions
No longer support vectors.  This change isn't backward compatible but
I haven't seen a vector being used even just once.

It also makes sense to take inspiration from `keymap-set' et al., which
didn't exist when Transient was created.

However, we continue to be more permissive than those functions, only
insisting on strings understood by `kbd', including strings that do not
satisfy `key-valid-p'.  Being more permissive makes it possible, for
example, to write the key binding, which toggles the "-a" command line
argument, as "-a", instead of having to write "- a".  Likewise
additional spaces can be added, which are not removed when displaying
the binding in the menu, which is useful for alignment purposes.