transient-cons-option: New suffix class for non-cmdline options
The existing `transient-option' class is suitable for options passed to
commands.  A suffix that uses that class may represent an argument such
as "--option".  In the list returned by `transient-args', that argument
and its value, "value", is represented as one element "--option=value".

For "options" that are not passed to a command that is not appropriate.
A caller that wants to extract the value of such an option would have
to parse that string.  `transient-arg-value' can help with that, but it
always returns a string (or nil).

The new `transient-cons-option' class provides a more direct way to deal
with "non-command-line options", by representing the key value pair
using a cons-cell (ARGUMENT . VALUE), not just internally, but in
particular when handing it of to a caller.  VALUE can have any printable
type and ARGUMENT can have any printable type, expect string.

We cannot use the term "key" because that is already used to the key
binding for the command.  Also avoid using "property" because that
implies a property list, while the value returned by `transient-args'
is an alist when all infixes use the `transient-cons-options' class.
Sticking to "argument" also has the advantage, that we don't have to
implement or generalize a dozen methods.