Inhibit quitting while a menu is active
Like the somewhat similar but much less complex `read-key-sequence',
for example, Transient menus should inhibit quitting, otherwise C-g
might, for example, kill some subprocess, when the user intended to
exit the menu.

Due to how Transient is implemented, there is nothing around which we
could let-bind `inhibit-quit', like that is usually done.  Instead we
have to set the global value and make sure we reset that when exiting
the menu.

The resetting has to happen on `post-command-hook' but at that time
a let-binding for `inhibit-quit' is in effect, so we have to use
`set-default-toplevel-value', instead of `setq'.

Closes #388.