transient--emergency-exit: Always reset inhibit-quit
We already call `transient--post-exit', which also takes care of this,
but to be absolutely sure, also reset here and unconditionally (i.e.,
even if `transient--prefix' is nil or `transient--exitp' is `replace').

Before the emergency exit was added, unexpected bugs could look up
Emacs.  That does not happen anymore, and we can also rely on it
to guarantee quitting is enabled again.