transient--init-suffix: Do not use :level twice calling constructor
For class constructors and `clone', if a keyword argument is specified
multiple times, the latter instance wins, which in our case was the
outdated one.

As a result, it was impossible to change the level of a suffix if that
was set inline, in the prefix definition.

Fixes [1: bcc0bf83].

1: 2025-05-16 bcc0bf836808e156262d9ff631be7d91a0ec201d
   Improve format of layout and handling of included groups