Improve kludge to work around an Emacs quitting kludge
Emacs commit 60c10b40d5b addresses this issued.  That hasn't landed
on master yet, so let's copy the message here:

> Make C-g in command loops consistently invoke quit keybinding
>
> C-g in command loops could unpredictably either signal an error or
> invoke the quit_char keybinding, depending on the relative timing
> of input and redisplay.  This change ensures C-g always runs the
> quit_char keybinding by wrapping read_char to check for quit_flag
> after reading input.

For older Emacs versions we have to work around the issue using this
new advice, approximately as suggested by Daniel Colascione at
https://github.com/magit/transient/commit/45fbefdc5b112f0a15cd93657.

The `unreadp' variable is an additional safety net, which should not
be necessary.  It would kick in if the unread C-g again signaled an
error, or if somehow our transient keymap got disabled, without this
this advice also getting disabled.

Also see #388.