transient--resume-override: No longer call transient--show
It is unnecessary to call `transient--show' here because it will
later happen anyway when `transient--post-command' calls either
`transient--refresh-transient' or `transient--redisplay'.

Calling `transient--show' here also meant that it could not take any
state changes into account that the invoked command made, which is
relevant if `refresh-suffixes' is non-nil and those changes effect
the result of, e.g., suffix predicates.

Closes #397.