transient--insert-suffix: Handle duplicated insertion using coordinates
When the suffix being inserted is already part of the menu, we
remove the existing entry and then insert a new entry.  In some
cases we could avoid these two steps and just do nothing, but it
would likely be more expensive and certainly more complicated to
determine whether we can skip the two steps than it is to just
do them.

When the same suffix is added twice at the same coordinates (e.g.,
because the same form ends up being evaluated twice), this resulted
in an error because the entry that is removed in this case also is
the pointer to the insertion location, and removing it invalidates
it.  So we have to lookup the coordinates again, after the removal.

Closes #400.
Closes #401.

(We cannot use `pcase-setq' because that wasn't added until Emacs
28.1 and Compat does not backport it.  Also, it turns out this is
the exception to my rule to never set more than one variable using
a single `setq' form.)

Co-authored-by: Evgeny Boykov <artscan@list.ru>